check if registration email exists

+++ resources/js/Pages/Auth/Vendor.vue
@@ -290,6 +290,18 @@ export default {
         // error: String,
     },
     computed: {
+        emailError(){
+            return this.$page.props.flash.error
+        }    
+    },
+    watch: {
+      emailError(newValue){
+          if(newValue == 'Email Address already Registered'){
+              this.taken()
+          }
+      }  
+    },
+    created () {
         
     },
     data () {
@@ -308,15 +320,30 @@ export default {
         }
     },
     methods: {
+        taken(){
+            Swal.fire(
+                    'Registration Error',
+                    'The email address has already been registered',
+                    'question'
+                ) 
+        },
         register(){
-            if( this.form.password == this.form.confirmPass ){
-                this.passError = ''
-                console.log(Ziggy.routes)
-                this.form.post(this.route('registerAdmin'))
-                localStorage.setItem('email', this.form.email)
-            }else{
-                this.passError = 'Passwords do not match'
-            }
+
+            if(this.emailError == 'Email Address already Registered'){
+                this.taken() 
+            }       
+            else{
+                if( this.form.password == this.form.confirmPass ){
+                    this.passError = ''
+                    // console.log(Ziggy.routes)
+                    this.form.post(this.route('registerAdmin'))
+                    localStorage.setItem('email', this.form.email)
+                }else{
+                    this.passError = 'Passwords do not match'
+                }
+            } 
+
+
             // console.log(this.error)
         }
     }
